# Added TimestampRulesFilter implementation