--- a/Sources/WhisperKit/Core/Models.swift
+++ b/Sources/WhisperKit/Core/Models.swift
@@ -191,9 +191,10 @@ public struct DecodingCache {
 ///   - sampleLength: The maximum number of tokens to sample.
 ///   - topK: Number of candidates when sampling with non-zero temperature.
 ///   - usePrefillPrompt: If true, the prefill tokens will be forced according to task and language settings.
-///   - usePrefillPrompt: If true, the kv cache will be prefilled based on the prefill data mlmodel.
+///   - usePrefillCache: If true, the kv cache will be prefilled based on the prefill data mlmodel.
 ///   - skipSpecialTokens: Whether to skip special tokens in the output.
 ///   - withoutTimestamps: Whether to include timestamps in the transcription result.
+///   - maxInitialTimestamp: Maximal initial timestamp.
 ///   - suppressBlank: If true, blank tokens will be suppressed during decoding.
 ///   - supressTokens: List of token IDs to suppress during decoding.
 ///   - compressionRatioThreshold: If the compression ratio of the transcription text is above this value, it is too repetitive and treated as failed.
@@ -215,6 +216,7 @@ public struct DecodingOptions {
     public var skipSpecialTokens: Bool
     public var withoutTimestamps: Bool
     public var wordTimestamps: Bool
+    public var maxInitialTimestamp: Float?
     public var clipTimestamps: [Float]
     public var suppressBlank: Bool
     public var supressTokens: [Int]
@@ -235,6 +237,7 @@ public struct DecodingOptions {
                 skipSpecialTokens: Bool = false,
                 withoutTimestamps: Bool = false,
                 wordTimestamps: Bool = false,
+                maxInitialTimestamp: Float? = nil,
                 clipTimestamps: [Float] = [],
                 suppressBlank: Bool = false,
                 supressTokens: [Int]? = nil,
@@ -255,6 +258,7 @@ public struct DecodingOptions {
         self.skipSpecialTokens = skipSpecialTokens
         self.withoutTimestamps = withoutTimestamps
         self.wordTimestamps = wordTimestamps
+        self.maxInitialTimestamp = maxInitialTimestamp
         self.clipTimestamps = clipTimestamps
         self.suppressBlank = suppressBlank
         self.supressTokens = supressTokens ?? [] // nonSpeechTokens() // TODO: implement these as default