--- a/Sources/WhisperKit/Core/Utils.swift
+++ b/Sources/WhisperKit/Core/Utils.swift
@@ -199,7 +199,7 @@ public func modelSupport(for deviceName: String) -> (default: String, disabled: 
             return ("base", ["small", "small.en", "large-v3_turbo", "large-v3", "large-v3_turbo_1018MB", "large-v2", "large-v2_turbo", "large-v2_turbo_1022MB", "large-v2_1050MB"])
 
         case let model where model.hasPrefix("iPhone13"): // A14
-            return ("base", ["large-v3_turbo", "large-v3", "large-v3_turbo_1018MB", "large-v2", "large-v2_turbo", "large-v2_turbo_1022MB"])
+            return ("base", ["large-v3_turbo", "large-v3", "large-v3_turbo_1018MB", "large-v2", "large-v2_1050MB", "large-v2_turbo", "large-v2_turbo_1022MB"])
 
         case let model where model.hasPrefix("iPhone14"), // A15
              let model where model.hasPrefix("iPhone15"), // A16