… Sources/WhisperKitCLI/transcribe.swift
… Tests/WhisperKitTests/UnitTests.swift