This PR adds implementation for `TimestampRulesFilter`. The implementation is based on https://github.com/openai/whisper/blob/master/whisper/decoding.py#L441

Couple of questions here @ZachNagengast:
- `sampleBegin` param passed to `TimestampRulesFilter` is 0, I think it might be incorrect. I compared it to the python implementation from the OpenAI repo and there this param is always greater or equal than 3 (and this makes sense, first 3 tokens are special tokens: 50258, 50259 and 50359 and AFAIK we don't want to supress them). If you run this code as is, some segments might be omited (because of the `sampleBegin` is 0, if you change it to 3, it should be ok).
- this implementation slows down the whole inference code, maybe you have some ideas how to optimize it?
- you mentioned that is has duplicated logic with `SegmentSeeker`, but I don't see it (AFAIK `TimestampRulesFilter` just supresses the token probabilities, while `SegmentSeeker` creates the whole segments). Could you please clarify?