@ZachNagengast 2024-03-13
@jkrukowski I push a small commit to measure the logit filtering time, here is what I'm getting for tiny with and without these new timestamp rules on the `jfk.wav` file:
With:
`[WhisperKit] - Logit Filtering:     192.41 ms /     28 runs (    6.87 ms/run) 37.78%`
Without:
`[WhisperKit] - Logit Filtering:       0.07 ms /     28 runs (    0.00 ms/run)  0.02%`


This is a bit high, it becomes especially noticeable with the tiny model. Something that is interesting is that only the first and last few tokens are slow (graph by chatgpt). This is for the jfk.wav

![image](https://github.com/argmaxinc/WhisperKit/assets/1981179/bd293b57-bd32-4a3f-9f7c-1d8422f420e4)
Hopefully this gives you some guidance on where to look for optimizations. And the majority of the slowdown is in this block of code:

```swift
            // timestamps have to appear in pairs, except directly before EOT; mask logits accordingly
            let sampledTokens = tokens[sampleBegin...]
            let lastWasTimestamp = sampledTokens.count >= 1 && sampledTokens.last! >= timeTokenBegin
            let penultimateWasTimestamp = sampledTokens.count < 2 || sampledTokens.dropLast().last! >= timeTokenBegin
            if lastWasTimestamp {
                if penultimateWasTimestamp {
                    // has to be non-timestamp
                    logits.fillLastDimension(indexes: timeTokenBegin..<logits.count, with: -FloatType.infinity)
                } else {
                    // cannot be normal text tokens
                    logits.fillLastDimension(indexes: 0..<endToken, with: -FloatType.infinity)
                }
            }
```