@jkrukowski 2024-03-19
> @jkrukowski I push a small commit to measure the logit filtering time, here is what I'm getting for tiny with and without these new timestamp rules on the `jfk.wav` file: With: `[WhisperKit] - Logit Filtering: 192.41 ms / 28 runs ( 6.87 ms/run) 37.78%` Without: `[WhisperKit] - Logit Filtering: 0.07 ms / 28 runs ( 0.00 ms/run) 0.02%`
> 
> This is a bit high, it becomes especially noticeable with the tiny model. Something that is interesting is that only the first and last few tokens are slow (graph by chatgpt). This is for the jfk.wav
> 
> ![image](https://private-user-images.githubusercontent.com/1981179/312273247-bd293b57-bd32-4a3f-9f7c-1d8422f420e4.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MTA4NjU1MzYsIm5iZiI6MTcxMDg2NTIzNiwicGF0aCI6Ii8xOTgxMTc5LzMxMjI3MzI0Ny1iZDI5M2I1Ny1iZDMyLTRhM2YtOWY3Yy0xZDg0MjJmNDIwZTQucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI0MDMxOSUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNDAzMTlUMTYyMDM2WiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9NmY1MTM4ZWQ0ZDM0NzYzNGIzMzJiNTBiMDNiNWU1MWM1ZWQwNGNlNTFkYTExNDdiNzMzNTM5MWI3YTVmZTliMSZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QmYWN0b3JfaWQ9MCZrZXlfaWQ9MCZyZXBvX2lkPTAifQ.8N37x7PObbfV92iulsOCcFrJpIlnSElRfEmI19DYVd0) Hopefully this gives you some guidance on where to look for optimizations. And the majority of the slowdown is in this block of code:
> 
> ```swift
>             // timestamps have to appear in pairs, except directly before EOT; mask logits accordingly
>             let sampledTokens = tokens[sampleBegin...]
>             let lastWasTimestamp = sampledTokens.count >= 1 && sampledTokens.last! >= timeTokenBegin
>             let penultimateWasTimestamp = sampledTokens.count < 2 || sampledTokens.dropLast().last! >= timeTokenBegin
>             if lastWasTimestamp {
>                 if penultimateWasTimestamp {
>                     // has to be non-timestamp
>                     logits.fillLastDimension(indexes: timeTokenBegin..<logits.count, with: -FloatType.infinity)
>                 } else {
>                     // cannot be normal text tokens
>                     logits.fillLastDimension(indexes: 0..<endToken, with: -FloatType.infinity)
>                 }
>             }
> ```

@ZachNagengast I've added more performant version of `fillLastDimension` function, seems like it's doing better, this is what I get for the release build on the `jfk.wav` file:

```
[WhisperKit] ---- Transcription Timings ----
[WhisperKit] Audio Load:              2.33 ms /      1 runs (    2.33 ms/run)  0.66%
[WhisperKit] Audio Processing:        0.11 ms /      1 runs (    0.11 ms/run)  0.03%
[WhisperKit] Mels:                   35.53 ms /      1 runs (   35.53 ms/run) 10.11%
[WhisperKit] Encoding:               13.39 ms /      1 runs (   13.39 ms/run)  3.81%
[WhisperKit] Matrices Init:           0.22 ms /      1 runs (    0.22 ms/run)  0.06%
[WhisperKit] Prefill:                 0.00 ms /      1 runs (    0.00 ms/run)  0.00%
[WhisperKit] Decoding:              239.40 ms /     28 runs (    8.55 ms/run) 68.15%
[WhisperKit] Non-inference:          61.25 ms /     28 runs (    2.19 ms/run) 17.43%
[WhisperKit] - Logit Filtering:       3.24 ms /     28 runs (    0.12 ms/run)  0.92%
[WhisperKit] - Sampling:             14.17 ms /     28 runs (    0.51 ms/run)  4.03%
[WhisperKit] - Kv Caching:            2.79 ms /     28 runs (    0.10 ms/run)  0.80%
[WhisperKit] - Word Timestamps:       0.00 ms /      0 runs (    0.00 ms/run)  0.00%
[WhisperKit] - Windowing:             0.08 ms /      1 runs (    0.08 ms/run)  0.02%
[WhisperKit] Fallbacks:               0.00 ms /      0 runs (    0.00 ms/run)  0.00%
[WhisperKit] Decoding Full Loop:    351.06 ms /     28 runs (   12.54 ms/run) 99.93%
```