@ZachNagengast 2024-03-19
Much better! This looks in line with what I was seeing for those faster middle tokens previously. Think this is ready to come out of draft now?