@jkrukowski 2024-03-19
> Much better! This looks in line with what I was seeing for those faster middle tokens previously. Think this is ready to come out of draft now?

good to hear this, 2 things are left:
1. `self.sampleBegin = 3 // FIXME: it should not be hardcoded value` -- not sure what value should I put there
2. force unwrapping in `sumOfProbabilityOverTimestampsIsAboveAnyOtherToken` maybe we should not force unwrap and return false gracefully, wdyt?